Update test05.m
Yet another curve coded manually. In the near future, I will migrate the actual animations to a separate storage while keeping this repository for the code only. Believe it or not, some of the examples done so far test some challenging cases, thus forcing me to review the code and fix bugs.