Double double added
This example is kinda weird: 2 points on an hypercycloid and 2 points on an epicycloid, coordinated. I think it looks nice. RIght now it assumess that the inner wheel follows a path that is a linear scaling of the path followed by the outer wheel: that only happens for circles. The drawing algorithm needs to be adjusted for general curves.